Clean up

+++ api/views/file_crud.py
@@ -18,6 +18,7 @@ router = InferringRouter(tags=['Files'])
 class FileCRUD:
     """This class based view deals with file and folder management.
     files can only be used in root folder 'file_storage' - for security reasons."""
+
     ROOT_FOLDER = 'file_storage/'
 
     @router.post("/files/{folder_path:path}", status_code=status.HTTP_201_CREATED)
@@ -25,7 +26,9 @@ class FileCRUD:
                               file: UploadFile = File(...)) -> APIMessage:
         """Uploads a file given a path"""
 
-        file_path: FilePath = self.get_and_validate_file_path(str(folder_path), exist=True, file_name=file.filename)
+        file_path: FilePath = self.get_and_validate_file_path(folder_path, file_name=file.filename)
+        if file_path.exists():
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"File {file_path} already exist.")
         try:
             await self.write_file(file_path, file)
             return APIMessage(detail=f"File {file.filename} created at {folder_path}")
@@ -36,15 +39,21 @@ class FileCRUD:
     def get_file(self, file_path: str):
         """Retrieves a file given a path - if the file is empty is dose not return a download link"""
 
-        file_path: FilePath = self.get_and_validate_file_path(file_path, exist=False)
+        file_path = self.get_and_validate_file_path(file_path)
+        if not file_path.exists():
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"File {file_path} dose not exist.")
+
         return FileResponse(file_path, filename=file_path.name)
 
     @router.get("/folder/{folder_path:path}")
     def get_folder_list(self, folder_path: str) -> List[str]:
-        """Retrieves a the content of a folder given a path.\n
+        """Retrieves a the content of a folder given a path.
         For path root folder enter backslash - '/' """
 
-        folder_path: DirectoryPath = self.get_and_validate_file_path(folder_path, exist=True)
+        folder_path: DirectoryPath = self.get_and_validate_file_path(folder_path)
+        if not folder_path.exists():
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"Folder {folder_path} dose not exist.")
+
         return os.listdir(folder_path)
 
     @router.put("/files/{file_path:path}")
@@ -52,10 +61,12 @@ class FileCRUD:
         """Updates a existing file given a path"""
 
         try:
-            file_path = self.get_and_validate_file_path(file_path, exist=False, file_name=file.filename)
+            file_path = self.get_and_validate_file_path(file_path, file_name=file.filename)
+            if not file_path.exists():
+                raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"File {file_path} dose not exist.")
             await self.write_file(file_path, file)
-            return APIMessage(detail=f"File {file.filename} created!")
 
+            return APIMessage(detail=f"File {file.filename} created!")
         except OSError:
             raise HTTPException(detail=f"Failed updating the file {file_path.name}", status_code=status.HTTP_417_EXPECTATION_FAILED)
 
@@ -64,19 +75,20 @@ class FileCRUD:
         """Removes a existing file given a path, Can only remove files"""
 
         try:
-            file_path: FilePath = self.get_and_validate_file_path(file_path, exist=True)
+            file_path: FilePath = self.get_and_validate_file_path(file_path)
+            if not file_path.exists():
+                raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"File {file_path} dose not exist.")
             os.remove(file_path)
-            return APIMessage(detail=f"Deleted file {file_path.name}")
 
+            return APIMessage(detail=f"Deleted file {file_path.name}")
         except PermissionError:
             raise HTTPException(detail=f"File {file_path.name} is protected", status_code=status.HTTP_403_FORBIDDEN)
 
-    def get_and_validate_file_path(self, path: str, exist: bool, file_name: Optional[str] = '') -> Path:
+    def get_and_validate_file_path(self, path: str, file_name: Optional[str] = '') -> Path:
         if path.startswith('.'):  # preventing from accessing parents files.
             raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"Path cannot start with dot.")
 
         path = Path(self.ROOT_FOLDER + path + file_name)
-        self.check_file_exist(path, exist)
         return path
 
     @staticmethod
@@ -85,11 +97,3 @@ class FileCRUD:
             content = await file.read()
             await out_file.write(content)
 
-    @staticmethod
-    def check_file_exist(file_path: Path, exist: bool) -> bool:
-        """This checks if file exist or not depending on exist parameter."""
-        if exist and file_path.exists():
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"File {file_path} already exist.")
-        elif not (exist and file_path.exists()):
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"File {file_path} dose not exist.")
-        return True
